envkeep/pyproject.toml
[project]
name = "envkeep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.5",
    "typer>=0.16.0",
    "python-dotenv>=1.0.0",
]
